moving `AffiliateCredit` component to dedicated directory
