+++ resources/assets/components/AffiliateCredit/AffiliateCredit.js
